dom/security/test/csp/test_sendbeacon.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1234813 - sendBeacon should not throw if blocked by Content Policy</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<iframe style="width:100%;" id="testframe" src="file_sendbeacon.html"></iframe>

<script class="testbody" type="text/javascript">

/* Description of the test:
 * Let's try to fire a sendBeacon which gets blocked by CSP. Let's make sure
 * sendBeacon does not throw an exception.
 */
SimpleTest.waitForExplicitFinish();

// a postMessage handler used to bubble up the
// result from within the iframe.
window.addEventListener("message", receiveMessage, false);
function receiveMessage(event) {
  var result = event.data.result;
  is(result, "blocked-beacon-does-not-throw", "sendBeacon should not throw");
  window.removeEventListener("message", receiveMessage, false);
  SimpleTest.finish();
}

</script>
</body>
</html>
